+++ spring_app/src/main/java/tarea/spring_app/repository/NotaRepository.java
@@ -6,10 +6,10 @@ import tarea.spring_app.model.Nota;
 
 import java.util.List;
 
-public interface NotaRepository extends JpaRepository<Nota, Long> {
-    List<Nota> findByActividadId(Long actividadId);
+public interface NotaRepository extends JpaRepository<Nota, Integer> {
+    List<Nota> findByActividadId(Integer actividadId);
 
     @Query("SELECT AVG(n.nota) FROM Nota n WHERE n.actividadId = ?1")
-    Double getPromedioByActividadId(Long actividadId);
+    Double getPromedioByActividadId(Integer actividadId);
 
 }
